cloud_functions/extract/requirements.txt
gcsfs
stravalib
pandas
google-cloud-secret-manager
requests
google-cloud-storage
pyarrow
google-cloud-bigquery 